+++ apps/publish/src/make/insertEntities.ts
@@ -15,6 +15,7 @@ import { EntityS3Writer } from "../entities/EntityS3Writer.js"
 import { cleanEntitiesS3 } from "../entities/cleanEntitiesS3.js"
 import { cleanTables } from "./cleanEntities.js"
 import getContributorJSON from "./getContributorJSON.js"
+import getAuthorizedNamespaces from "./getAuthorizedNamespaces.js"
 import getImageJSON from "./getImageJSON.js"
 import getNodeJSON from "./getNodeJSON.js"
 import type { SourceData } from "./getSourceData.js"
@@ -293,6 +294,13 @@ const insertEntities = async (client: ClientBase, data: SourceData, isDryRun = f
     await insertIllustrations(client, data, isDryRun)
     await client.query(isDryRun ? "ROLLBACK" : "COMMIT")
     if (s3Writer) {
+        s3Writer.putStatic(
+            "namespaces",
+            stringifyNormalized({
+                build: data.build,
+                namespaces: getAuthorizedNamespaces(data),
+            }),
+        )
         console.info("Uploading entity JSON to S3...")
         await s3Writer.flush()
         console.info("Uploaded entity JSON to S3.")